🧩  Move to dg layout
- Updated Makefile to streamline commands and ensure compatibility with `uv`.
- Upgraded dependencies in `pyproject.toml` for `dagster`, `dagster-dg`, and `polars`.
- Added `.gitignore` entry to include the `lib` directory.
- Removed unused files related to Hugging Face and indicators assets, definitions, and resources.
