+++ pyproject.toml
@@ -3,15 +3,12 @@ name = "datadex"
 version = "1.0.0"
 authors = [{ name = "David Gasquez" }]
 dependencies = [
-    "dagster>=1.10.9",
-    "dagster-duckdb>=0.26.9",
-    "dagster-duckdb-polars>=0.26.9",
-    "dagster-webserver>=1.10.9",
-    "datasets>=3.5.0",
-    "duckdb>=1.2.2",
+    "dagster>=1.10.13",
+    "dagster-dg>=0.26.13",
+    "dagster-duckdb-polars>=0.26.13",
     "httpx>=0.28.1",
     "huggingface-hub>=0.30.2",
-    "polars>=1.27.1",
+    "polars>=1.29.0",
 ]
 
 requires-python = ">=3.11, <3.14"
@@ -24,21 +21,23 @@ Changelog = "https://github.com/datonic/datadex/commits/main/"
 Issues = "https://github.com/datonic/datadex/issues"
 CI = "https://github.com/datonic/datadex/actions"
 
-[tool.dagster]
-module_name = "datadex.definitions"
-code_location_name = "datadex"
+[project.entry-points]
+"dagster_dg.plugin" = { datadex = "datadex.lib" }
 
 [dependency-groups]
 dev = [
-    "ipykernel>=6.29.5",
+    "dagster-webserver>=1.10.13",
 ]
 
 [build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
+requires = ["uv_build>=0.6.14,<0.7"]
+build-backend = "uv_build"
 
-[tool.hatch.build.targets.wheel]
-packages = ["datadex"]
+[tool.dg]
+directory_type = "project"
 
-[tool.pyright]
-reportUnknownMemberType = false
+[tool.dg.project]
+root_module = "datadex"
+
+[tool.dg.project.python_environment]
+active = true
